requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
ccxt==4.4.68
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dnspython==2.7.0
frozenlist==1.5.0
idna==3.10
motor==3.7.0
multidict==6.1.0
propcache==0.3.0
pycares==4.5.0
pycparser==2.22
pymongo==4.11.2
python-dotenv==1.0.1
requests==2.32.3
setuptools==76.0.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
